pyproject.toml
[tool.poetry]
name = "pyrogyro"
version = "0.0.1"
description = "Yet Another (Gyro) Gamepad Mapper, written in Python for hackability"
authors = ["ForOhForError"]
packages = [{include = "pyrogyro"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pysdl3 = "^0.9.1b2"
vgamepad = "^0.1.0"
infi-systray = "^0.1.12.1"
pyautogui = "^0.9.54"
flask = "^3.1.0"
flask-sock = "^0.7.0"

[tool.poetry.scripts]
pyrogyro = "pyrogyro.pyrogyro:appmain"
dist = "pyrogyro.installer:install_win"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.11.1"
black = "^24.10.0"
isort = "^5.13.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
